# api-contracts/documents/openapi.json
{"openapi":"3.1.0","info":{"title":"Documents API","description":"RESTful API for managing document sets, schemas, and document versions.\n\n## Features\n- Create and manage document sets with multiple document types\n- Version control for documents with immutable versions\n - Schema management with versioning support\n - Document validation against schemas\n - Derivative document creation (transformations)\n - Raw document-version content download, including preserved EDI source payloads\n - Content-addressable storage with hash verification\n\n## Domain Concepts\n- **Document Set**: Container for related documents of different types\n- **Document**: A single document within a set (e.g., invoice, order)\n- **Document Version**: Immutable snapshot of document content\n- **Schema**: Defines structure and validation rules for documents\n- **Derivative**: Transformed version of a document (e.g., PDF from XML)\n","contact":{"name":"API Support","email":"support@example.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0.html"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Local development server"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Schemas","description":"Manage schemas and schema versions for document validation"},{"name":"Document Sets","description":"Manage document sets, documents, versions, and derivatives"}],"paths":{"/api/schemas":{"post":{"tags":["Schemas"],"summary":"Create a new schema","description":"Creates a new schema for document validation. Schemas define structure and validation rules.","operationId":"createSchema","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSchemaRequest"}}},"required":true},"responses":{"201":{"description":"Schema created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SchemaResponse"}}}},"400":{"description":"Invalid request data","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}}}}},"/api/schemas/{schemaId}/versions":{"post":{"tags":["Schemas"],"summary":"Add a schema version","description":"Adds a new version to an existing schema. Schema definition must be Base64 encoded.","operationId":"addSchemaVersion","parameters":[{"name":"schemaId","in":"path","description":"Schema UUID","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddSchemaVersionRequest"}}},"required":true},"responses":{"201":{"description":"Schema version added successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SchemaVersionResponse"}}}},"400":{"description":"Invalid request data","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Schema not found","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}}}}},"/api/document-sets":{"get":{"tags":["Document Sets"],"summary":"List document sets","description":"Retrieves document sets with pagination. Use nextToken or nextUrl from response to fetch subsequent pages.","operationId":"listDocumentSets","parameters":[{"name":"limit","in":"query","description":"Page size (1-100, default 20)","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"nextToken","in":"query","description":"Continuation token from previous page","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Document sets retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginatedResponseDocumentSetResponse"}}}},"400":{"description":"Invalid limit or pagination token","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}}}},"post":{"tags":["Document Sets"],"summary":"Create a new document set","description":"Creates a new document set with an initial document. The document content must be Base64 encoded.","operationId":"createDocumentSet","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDocumentSetRequest"}}},"required":true},"responses":{"201":{"description":"Document set created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentSetResponse"}}}},"400":{"description":"Invalid request data","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Schema not found","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}}}}},"/api/document-sets/{setId}/documents":{"post":{"tags":["Document Sets"],"summary":"Add a document to a set","description":"Adds a new document to an existing document set. Content must be Base64 encoded.","operationId":"addDocument","parameters":[{"name":"setId","in":"path","description":"Document set UUID","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddDocumentRequest"}}},"required":true},"responses":{"201":{"description":"Document added successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentResponse"}}}},"400":{"description":"Invalid request data","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Document set or schema not found","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}}}}},"/api/document-sets/{setId}/documents/{docId}/versions":{"post":{"tags":["Document Sets"],"summary":"Add a document version","description":"Adds a new version to an existing document. Content must be Base64 encoded.","operationId":"addDocumentVersion","parameters":[{"name":"setId","in":"path","description":"Document set UUID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"docId","in":"path","description":"Document UUID","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddVersionRequest"}}},"required":true},"responses":{"201":{"description":"Version added successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentVersionResponse"}}}},"400":{"description":"Invalid request data","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Document set or document not found","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}}}}},"/api/document-sets/{setId}/documents/{docId}/versions/{versionNumber}/validate":{"post":{"tags":["Document Sets"],"summary":"Validate a document version","description":"Validates a document version against its schema. Returns 200 if valid, 422 if invalid.","operationId":"validateDocumentVersion","parameters":[{"name":"setId","in":"path","description":"Document set UUID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"docId","in":"path","description":"Document UUID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"versionNumber","in":"path","description":"Version number (1-based)","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Document is valid","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ValidationResultResponse"}}}},"422":{"description":"Document is invalid","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Document set, document, or version not found","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}}}}},"/api/document-sets/{setId}/documents/{docId}/derivatives":{"get":{"tags":["Document Sets"],"summary":"Get all derivatives","description":"Retrieves all derivatives for a document","operationId":"listDerivatives","parameters":[{"name":"setId","in":"path","description":"Document set UUID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"docId","in":"path","description":"Document UUID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Derivatives retrieved successfully","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DerivativeResponse"}}}}},"404":{"description":"Document set or document not found","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}}}},"post":{"tags":["Document Sets"],"summary":"Create a derivative","description":"Creates a derivative (transformed version) from a document version in a different format","operationId":"createDerivative","parameters":[{"name":"setId","in":"path","description":"Document set UUID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"docId","in":"path","description":"Document UUID","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDerivativeRequest"}}},"required":true},"responses":{"201":{"description":"Derivative created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DerivativeResponse"}}}},"400":{"description":"Invalid request data or duplicate derivative","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Document set, document, or version not found","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}}}}},"/api/schemas/{schemaId}/versions/{versionId}":{"get":{"tags":["Schemas"],"summary":"Get a schema version","description":"Retrieves a specific version of a schema","operationId":"getSchemaVersion","parameters":[{"name":"schemaId","in":"path","description":"Schema UUID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"versionId","in":"path","description":"Version identifier (e.g., '1.0', '2.1')","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Schema version found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SchemaVersionResponse"}}}},"404":{"description":"Schema or schema version not found","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}}}}},"/api/schemas/{id}":{"get":{"tags":["Schemas"],"summary":"Get a schema","description":"Retrieves a schema by its unique identifier","operationId":"getSchema","parameters":[{"name":"id","in":"path","description":"Schema UUID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Schema found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SchemaResponse"}}}},"404":{"description":"Schema not found","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}}}}},"/api/document-sets/{setId}/documents/{docId}":{"get":{"tags":["Document Sets"],"summary":"Get a document","description":"Retrieves a document by its unique identifier within a document set","operationId":"getDocument","parameters":[{"name":"setId","in":"path","description":"Document set UUID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"docId","in":"path","description":"Document UUID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Document found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentResponse"}}}},"404":{"description":"Document set or document not found","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}}}}},"/api/document-sets/{setId}/documents/{docId}/versions/{versionNumber}":{"get":{"tags":["Document Sets"],"summary":"Get a document version","description":"Retrieves a specific version of a document by version number","operationId":"getDocumentVersion","parameters":[{"name":"setId","in":"path","description":"Document set UUID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"docId","in":"path","description":"Document UUID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"versionNumber","in":"path","description":"Version number (1-based)","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Version found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentVersionResponse"}}}},"404":{"description":"Document set, document, or version not found","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}}}}},"/api/document-sets/{setId}/documents/{docId}/versions/{versionNumber}/content":{"get":{"tags":["Document Sets"],"summary":"Get document version content","description":"Retrieves the raw content bytes for a specific document version","operationId":"getDocumentVersionContent","parameters":[{"name":"setId","in":"path","description":"Document set UUID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"docId","in":"path","description":"Document UUID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"versionNumber","in":"path","description":"Version number (1-based)","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Version content found","headers":{"X-Document-Set-Id":{"description":"Document set identifier","style":"simple"},"X-Document-Version":{"description":"Document version number","style":"simple"},"X-Content-Hash":{"description":"Stored content hash","style":"simple"},"X-Document-Id":{"description":"Document identifier","style":"simple"},"X-Document-Format":{"description":"Stored document format","style":"simple"}},"content":{"application/octet-stream":{"schema":{"type":"string","format":"binary"}}}},"404":{"description":"Document set, document, version, or content not found","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}}}}},"/api/document-sets/{id}":{"get":{"tags":["Document Sets"],"summary":"Get a document set","description":"Retrieves a document set by its unique identifier","operationId":"getDocumentSet","parameters":[{"name":"id","in":"path","description":"Document set UUID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Document set found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentSetResponse"}}}},"404":{"description":"Document set not found","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/ApiError"}}}}}}}},"components":{"schemas":{"CreateSchemaRequest":{"type":"object","properties":{"name":{"type":"string","minLength":1},"format":{"type":"string","enum":["XSD","JSON_SCHEMA","RELAXNG"]}},"required":["format","name"]},"SchemaResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"format":{"type":"string","enum":["XSD","JSON_SCHEMA","RELAXNG"]},"versions":{"type":"array","items":{"$ref":"#/components/schemas/SchemaVersionSummary"}}}},"SchemaVersionSummary":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"versionIdentifier":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"deprecated":{"type":"boolean"}}},"AddSchemaVersionRequest":{"type":"object","properties":{"versionIdentifier":{"type":"string","minLength":1},"definition":{"type":"string","minLength":1}},"required":["definition","versionIdentifier"]},"SchemaVersionResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"versionIdentifier":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"deprecated":{"type":"boolean"}}},"CreateDocumentSetRequest":{"type":"object","properties":{"documentType":{"type":"string","enum":["INVOICE","ORDER","CREDIT_NOTE","DEBIT_NOTE","QUOTATION","DESPATCH_ADVICE","RECEIPT_ADVICE","STATEMENT","REMITTANCE_ADVICE","CATALOGUE","APPLICATION_RESPONSE"]},"schemaId":{"type":"string","format":"uuid"},"schemaVersion":{"type":"string","minLength":1},"content":{"type":"string","minLength":1},"createdBy":{"type":"string","minLength":1},"metadata":{"type":"object","additionalProperties":{"type":"string"}}},"required":["content","createdBy","documentType","schemaId","schemaVersion"]},"DocumentSetResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"metadata":{"type":"object","additionalProperties":{"type":"string"}},"documents":{"type":"array","items":{"$ref":"#/components/schemas/DocumentSummary"}}}},"DocumentSummary":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"type":{"type":"string","enum":["INVOICE","ORDER","CREDIT_NOTE","DEBIT_NOTE","QUOTATION","DESPATCH_ADVICE","RECEIPT_ADVICE","STATEMENT","REMITTANCE_ADVICE","CATALOGUE","APPLICATION_RESPONSE"]},"versionCount":{"type":"integer","format":"int32"}}},"AddDocumentRequest":{"type":"object","properties":{"documentType":{"type":"string","enum":["INVOICE","ORDER","CREDIT_NOTE","DEBIT_NOTE","QUOTATION","DESPATCH_ADVICE","RECEIPT_ADVICE","STATEMENT","REMITTANCE_ADVICE","CATALOGUE","APPLICATION_RESPONSE"]},"schemaId":{"type":"string","format":"uuid"},"schemaVersion":{"type":"string","minLength":1},"content":{"type":"string","minLength":1},"createdBy":{"type":"string","minLength":1},"relatedDocumentId":{"type":"string","format":"uuid"}},"required":["content","createdBy","documentType","schemaId","schemaVersion"]},"DerivativeResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"sourceVersionId":{"type":"string","format":"uuid"},"targetFormat":{"type":"string","enum":["XML","JSON","PDF","EDI"]},"contentHash":{"type":"string"},"transformationMethod":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"DocumentResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"type":{"type":"string","enum":["INVOICE","ORDER","CREDIT_NOTE","DEBIT_NOTE","QUOTATION","DESPATCH_ADVICE","RECEIPT_ADVICE","STATEMENT","REMITTANCE_ADVICE","CATALOGUE","APPLICATION_RESPONSE"]},"schemaRef":{"$ref":"#/components/schemas/SchemaRefResponse"},"versionCount":{"type":"integer","format":"int32"},"currentVersion":{"$ref":"#/components/schemas/DocumentVersionResponse"},"derivatives":{"type":"array","items":{"$ref":"#/components/schemas/DerivativeResponse"}}}},"DocumentVersionResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"versionNumber":{"type":"integer","format":"int32"},"contentHash":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"format":{"type":"string","enum":["XML","JSON","PDF","EDI"]},"parseStatus":{"type":"string"},"messageType":{"type":"string"},"parseErrors":{"type":"array","items":{"type":"string"}}}},"SchemaRefResponse":{"type":"object","properties":{"schemaId":{"type":"string","format":"uuid"},"version":{"type":"string"}}},"AddVersionRequest":{"type":"object","properties":{"content":{"type":"string","minLength":1},"createdBy":{"type":"string","minLength":1}},"required":["content","createdBy"]},"ValidationErrorResponse":{"type":"object","properties":{"path":{"type":"string"},"message":{"type":"string"}}},"ValidationResultResponse":{"type":"object","properties":{"valid":{"type":"boolean"},"errors":{"type":"array","items":{"$ref":"#/components/schemas/ValidationErrorResponse"}},"warnings":{"type":"array","items":{"$ref":"#/components/schemas/ValidationWarningResponse"}}}},"ValidationWarningResponse":{"type":"object","properties":{"path":{"type":"string"},"message":{"type":"string"}}},"CreateDerivativeRequest":{"type":"object","properties":{"sourceVersionNumber":{"type":"integer","format":"int32","minimum":1},"targetFormat":{"type":"string","enum":["XML","JSON","PDF","EDI"]}},"required":["targetFormat"]},"PaginatedResponseDocumentSetResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/DocumentSetResponse"}},"pageSize":{"type":"integer","format":"int32"},"hasPrevious":{"type":"boolean"},"hasNext":{"type":"boolean"},"previousToken":{"type":"string"},"nextToken":{"type":"string"},"previousUrl":{"type":"string"},"nextUrl":{"type":"string"}}},"ApiError":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"details":{"type":"object","additionalProperties":{"type":"object"}}},"required":["code","message","timestamp","details"]}},"securitySchemes":{"bearerAuth":{"type":"http","description":"OIDC access token containing tenant_id and role claims","scheme":"bearer","bearerFormat":"JWT"}}}}